fix(axvisor): explicit tower util feature and rename http::axum to server
Address review items on the axum management plane.

Changes:
- os/axvisor/Cargo.toml: declare tower `util` explicitly. `ServiceExt::oneshot`
  is gated behind tower's `util` feature, which default features do not include;
  it only compiled via axum's transitive `util` enablement
- os/axvisor/src/http/mod.rs: rename the `http::axum` submodule to
  `http::server` to avoid shadowing the external axum crate (mod.rs
  `axum::serve()` previously resolved to the submodule while server.rs
  `axum::serve()` resolved to the crate); drop the redundant
  `#[cfg(feature = "http-axum")]` on module items, already gated by `mod http`
- os/axvisor/src/http/axum.rs -> os/axvisor/src/http/server.rs: file rename
- os/axvisor/src/http/vm.rs: update doc reference `[super::axum]` -> `[super::server]`
- os/axvisor/src/main.rs: correct the spawn-order comment. `ax_std::thread::spawn`
  only enqueues the task; the main task keeps running until it yields, so the
  HTTP server's bind is not guaranteed to precede `launch_default_vms`
