+++ os/axvisor/Cargo.toml
@@ -64,7 +64,7 @@ anyhow.workspace = true
 log = "0.4"
 axum = { version = "0.8", optional = true }
 serde_json = { version = "1", optional = true }
-tower = { version = "0.5", optional = true }
+tower = { version = "0.5", optional = true, features = ["util"] }
 tokio = { version = "1", optional = true, features = [
   "rt",
   "macros",
